requirements.txt
fastapi
uvicorn
pyserial
ultralytics
opencv-python